requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bottle==0.12.25
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clr-loader==0.2.5
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.6
proxy-tools==0.1.0
pycparser==2.21
pythonnet==3.0.1
pywebview==4.1
requests==2.30.0
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
